move Legend to seperate folder
