+++ src/components/Map/Legend/Legend.css
@@ -1,0 +1,25 @@
+.legend {
+  background: rgba(255, 255, 255, 0.8);
+  position: absolute;
+  bottom: 20px;
+  right: 0px;
+  float: right;
+  padding: 0.5rem;
+  margin: 0.5rem;
+  border: 0 black solid;
+  border-radius: 0.5rem;
+  display: flex;
+  flex-direction: column;
+}
+
+.legend .legend-color {
+  width: 10px;
+  height: 10px;
+  display: inline-block;
+  border-radius: 20%;
+  margin: 0.2rem 0.2rem 0 0;
+}
+
+.legend span {
+  text-transform: capitalize;
+}
